NPA-5235: Set Up Repository Documentation (#1)
<!-- markdownlint-disable-next-line first-line-heading -->
## Description

Key change:
- Added documentation to readme to describe the project

Minor Changes:
- Ran prettier on all files, going to add a PR check for prettier in
another PR
- Merge drawio and png files
- Updated Pull Request template to pass markdown checks
- Removed Go/TypeScript scripts/modules that are causing dependabot
warnings

## Context

<!-- Why is this change required? What problem does it solve? -->

This Pull Request updates the documentation of the repository to begin
personalising the repository template for IM1 PFS Auth.

## Type of changes

<!-- What types of changes does your code introduce? Put an `x` in all
the boxes that apply. -->

- [x] Refactoring (non-breaking change)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would change existing
functionality)
- [ ] Bug fix (non-breaking change which fixes an issue)

## Checklist

<!-- Go over all the following points, and put an `x` in all the boxes
that apply. -->

- [x] I am familiar with the [contributing
guidelines](../docs/CONTRIBUTING.md)
- [x] I have followed the code style of the project
- [ ] I have added tests to cover my changes
- [x] I have updated the documentation accordingly
- [ ] This PR is a result of pair or mob programming

---

## Sensitive Information Declaration

To ensure the utmost confidentiality and protect your and others
privacy, we kindly ask you to NOT including [PII (Personal Identifiable
Information) / PID (Personal Identifiable
Data)](https://digital.nhs.uk/data-and-information/keeping-data-safe-and-benefitting-the-public)
or any other sensitive data in this PR (Pull Request) and the codebase
changes. We will remove any PR that do contain any sensitive
information. We really appreciate your cooperation in this matter.

- [x] I confirm that neither PII/PID nor sensitive data are included in
this PR and the codebase changes.