NPA-5235: Refactor GitHub Workflows (#2)
<!-- markdownlint-disable MD041 -->

## Description

<!-- Describe your changes in detail. -->

Refactor GitHub Workflows by removing CICD/Pull request steps that may
not be needed by this project. They can be re-added in the future if
necessary.

> [!NOTE]  
> Once this PR has been merged all `Run Code Checks` and `CodeQL
Analysis` checks will be made required for merging into main.

## Context

<!-- Why is this change required? What problem does it solve? -->

To stop workflows from performing actions that haven't been set up by
this team's dev. So that the entire CICD workflows are under our control
and understanding.

## Type of changes

<!-- What types of changes does your code introduce? Put an `x` in all
the boxes that apply. -->

- [x] Refactoring (non-breaking change)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would change existing
functionality)
- [ ] Bug fix (non-breaking change which fixes an issue)

## Checklist

<!-- Go over all the following points, and put an `x` in all the boxes
that apply. -->

- [x] I am familiar with the [contributing
guidelines](../docs/CONTRIBUTING.md)
- [x] I have followed the code style of the project
- [ ] I have added tests to cover my changes
- [x] I have updated the documentation accordingly
- [ ] This PR is a result of pair or mob programming

---

## Sensitive Information Declaration

To ensure the utmost confidentiality and protect your and others
privacy, we kindly ask you to NOT including [PII (Personal Identifiable
Information) / PID (Personal Identifiable
Data)](https://digital.nhs.uk/data-and-information/keeping-data-safe-and-benefitting-the-public)
or any other sensitive data in this PR (Pull Request) and the codebase
changes. We will remove any PR that do contain any sensitive
information. We really appreciate your cooperation in this matter.

- [x] I confirm that neither PII/PID nor sensitive data are included in
this PR and the codebase changes.