NPA-5244: Update comments in e2e test files to be correct (#51)
# Pull Request

## 🧾 Ticket Link

https://nhsd-jira.digital.nhs.uk/browse/NPA-5244

---

## 📄 Description/Summary of Changes

<!-- Describe the changes made in this PR. Include the
purpose/scope/impact/context of the changes -->

- <!-- Add bullet points for changes made -->
Updated comments in the end to end testing files to be correct and
relate to the test expected behaviour
---

## 🧪 Developer Testing Carried Out

<!-- Describe what tests (automated/unit/manual etc.) have been done for
the ticket. Include: -->
<!-- - Any tests added/updated -->
<!-- - Evidence that each acceptance criterion from the Jira ticket is
met -->

- <!-- Add bullet points for tests done -->

---

## ✅ Developer Checklist

<!-- To be completed by the developer -->

- [ ] I have set the PR title to follow the format: `NPA-XXXX:
<short-description>`
- [ ] My branch name follows the convention:
`<type>/NPA-XXXX/<short-description>`
- [ ] My commit messages follow the template: `NPA-XXXX:
<short-description>`
- [ ] I have updated the documentation accordingly
- [ ] I have set assignees and added appropriate labels

---

## 👀 Reviewer Checklist

<!-- To be completed by the reviewer -->

- [ ] Changes meet the acceptance criteria of the Jira ticket
- [ ] Code is able to be merged (no conflicts and adheres to coding
standards)
- [ ] Sufficient test evidence is provided (manual and/or automated)

---

## Sensitive Information Declaration

To ensure the utmost confidentiality and protect your and others
privacy, we kindly ask you to NOT including [PII (Personal Identifiable
Information) / PID (Personal Identifiable
Data)](https://digital.nhs.uk/data-and-information/keeping-data-safe-and-benefitting-the-public)
or any other sensitive data in this PR (Pull Request) and the codebase
changes. We will remove any PR that do contain any sensitive
information. We really appreciate your cooperation in this matter.

- [ ] I confirm that neither PII/PID nor sensitive data are included in
this PR and the codebase changes.